+++ djpcms/apps/__init__.py
@@ -1,1 +1,7 @@
+'''While the :mod:`djpcms.core` defines some core functionalities
+and classes used throughtout the library,
+the :mod:`djpcms.apps` module represents the main
+interface and it is used to build real web-based
+dynamic applications.
+'''
 from .site import *