+++ djpcms/contrib/monitor/__init__.py
@@ -1,10 +1,17 @@
-'''
+'''\
 An utility for displaying admin panels of models and linking
-sql models with stdnet models
+sql models with stdnet models.
+It is useful for django_ based sites
+needing to use a *smart cache* or adding extra in memory data
+to their relational database models.
+
 
 Requires:
 
-* stdnet
-* django (optional)
+* stdnet_
+* django_ (optional)
+
+.. _stdnet: http://lsbardel.github.com/python-stdnet/
+.. _django: http://www.djangoproject.com/
 '''
-from .stdlink import *
+from stdlink import *